extract expect-ember-error helper
